Tests: Split integration and unit tests.
