+++ tests/vision/core.unit/syntheticBoard.cpp
