+++ tests/vision/core.unit/syntheticBoard.hpp
